Date format changed. Icon spacing.
Date locale is tricky to get; adding the "order by" clause in sql completely negates my efforts to get a french month out of the database. Nice. So I decided to leave it as a number. Added "fa-fw" class to font awesome icons in the experience section to give just a bit of spacing.